+++ src/vrml97/AudioClip.cpp
@@ -786,7 +786,7 @@ SoVRMLAudioClipP::internal_read(void * COIN_UNUSED_ARG(datasource), void *buffer
   while (!bufferFilled) {
     if (this->currentPause>0.0) {
       // deliver a zero'ed,  buffer
-      int outputsize = (numframes - framepos) * channelsdelivered * 
+      size_t outputsize = size_t(numframes - framepos) * size_t(channelsdelivered) *
         sizeof(int16_t);
       memset(((int16_t *)buffer) + framepos*channelsdelivered, 0, outputsize);
       this->currentPause -= (double)(numframes - framepos) / 
@@ -804,7 +804,7 @@ SoVRMLAudioClipP::internal_read(void * COIN_UNUSED_ARG(datasource), void *buffer
 
     if (this->playlist.getLength() == 0) {
       this->closeFile();
-      int outputsize = (numframes - framepos) * channelsdelivered * 
+      size_t outputsize = size_t(numframes - framepos) * size_t(channelsdelivered) *
         sizeof(int16_t);
       memset(((int16_t *)buffer) + framepos*channelsdelivered, 0, 
               outputsize);
@@ -824,7 +824,7 @@ SoVRMLAudioClipP::internal_read(void * COIN_UNUSED_ARG(datasource), void *buffer
         else {
           // We have played all files in the list, and we're not looping.
           // => We can stop playing.
-          int outputsize = (numframes - framepos) * channelsdelivered * 
+          size_t outputsize = size_t(numframes - framepos) * size_t(channelsdelivered) *
             sizeof(int16_t);
           memset(((int16_t *)buffer) + framepos*channelsdelivered, 0, 
                  outputsize);
